# docs/swagger.yaml
basePath: /api
definitions:
...
  controller.Setup2FARequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
...
  controller.Verify2FARequest:
    properties:
      code:
        type: string
      flow_token:
        type: string
    required:
    - code
    type: object
...
  dto.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact: {}
  description: new-api AI 网关 —— 用户与账号相关接口(对外第三方)
  title: new-api 用户 API
  version: "1.0"
paths:
...
  /user/2fa/backup_codes:
    post:
      parameters:
      - description: TOTP 验证码
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.Verify2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: 重新生成 2FA 备用码
      tags:
      - 用户-2FA
  /user/2fa/disable:
    post:
      parameters:
      - description: TOTP 验证码或备用码
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.Verify2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: 验证后禁用 2FA
      tags:
      - 用户-2FA
  /user/2fa/enable:
    post:
      parameters:
      - description: TOTP 验证码
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.Setup2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: 验证 TOTP 并启用 2FA
      tags:
      - 用户-2FA
  /user/2fa/setup:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: 初始化 2FA 设置，返回密钥/二维码/备用码
      tags:
      - 用户-2FA
  /user/2fa/status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
      security:
      - ApiKeyAuth: []
      summary: 查询当前用户 2FA 状态
      tags:
      - 用户-2FA
...
  /user/login/2fa:
    post:
      parameters:
      - description: 2FA 验证码(TOTP 或备用码)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller.Verify2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.APIResponse'
      summary: 登录二次验证(2FA)
      tags:
      - 用户-认证
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"